==> 1.ETL+EDA.ipynb <==
import json
import re
from textblob import TextBlob
import pandas as pd
import numpy as np

Rutas de los datasets

ruta_steam_games = 'datasets/output_steam_games.json'
ruta_users_items = 'datasets/australian_users_items.json'
ruta_users_reviews = 'datasets/australian_user_reviews.json'

ruta_parquet_games = "datasets/data_steam_games.parquet"
ruta_parquet_reviews = "datasets/data_reviews.parquet"
ruta_parquet_items = "datasets/data_items.parquet"

ruta_games_csv = 'datasets/steam_games.csv'
ruta_items_csv = 'datasets/users_items.csv'
ruta_reviews_csv = 'users_reviews.csv'

empezando extracción dataframe steam_games

dic_steam_games = []
with open(ruta_steam_games, 'r') as file:
    for line in file:
        json_data = json.loads(line)
        dic_steam_games.append(json_data)
df_games = pd.DataFrame(dic_steam_games)

df_games.head()

df_games.info()

Guardando el archivo como CSV

df_games.to_csv(ruta_games_csv, index=False, sep=';', encoding="utf-8")

df_games = pd.read_csv(ruta_games_csv, sep=";")

Removendo Líneas Duplicadas

df_games = df_games.drop_duplicates()

guardando el archivo como parquet

df_games.to_parquet(ruta_parquet_games)

empezando extracción dataframe user_items

with open(ruta_users_items, 'r', encoding='utf-8') as f:
    data =f.readlines()

records = [eval(line.strip()) for line in data]

df_users_items = pd.DataFrame(records)

# Crear el DataFrame con las columnas user_id, items_count, steam_id y user_url
df_users_info = df_users_items[['user_id', 'items_count', 'steam_id', 'user_url']]

# Crear una lista para almacenar los registros desglosados de items
items_records = []

# Recorrer el DataFrame original y desglosar los elementos de items
for index, row in df_users_items.iterrows():
    user_id = row['user_id']
    items_list = row['items']
    for item in items_list:
        item_record = {
            'user_id': user_id,
            'item_id': item['item_id'],
            'item_name': item['item_name'],
            'playtime_forever': item['playtime_forever'],
            # 'playtime_2weeks': item['playtime_2weeks']
        }
        items_records.append(item_record)

# Crear el DataFrame con los registros desglosados de items
df_items_info = pd.DataFrame(items_records)

df_items_info.head()

df_items_info.info()

guardando el archivo como csv

df_items_info.to_csv(ruta_items_csv, index=False, sep=';', encoding='utf-8')

df_items_info = pd.read_csv(ruta_items_csv,sep=";")

Removendo Líneas duplicadas

df_items_info = df_items_info.drop_duplicates()

guardando el archivo como parquet

df_items_info.to_parquet(ruta_parquet_items)

empezando extracción dataframe user_reviews

# Abre el archivo JSON y lee los datos
with open(ruta_users_reviews, 'r', encoding='utf-8') as f:
    data = f.readlines()
# Corrige el formato JSON eliminando comas al final de las líneas y encierra todo en corchetes para formar una lista de diccionarios válida
fixed_data = "[" + ",".join(data) + "]"
# Convierte el JSON en una lista de diccionarios
records = eval(fixed_data)
# Crea el DataFrame con las reviews y agrega la relación con user_id
df_users_reviews = pd.DataFrame([{
    "user_id": record["user_id"],
    "item_id": review["item_id"],
    "posted": review["posted"],
    "recommend": review["recommend"],
    "review": review["review"]
} for record in records for review in record["reviews"]])

df_users_reviews.head()

df_users_reviews.info()

guardando el archivo como csv

df_users_reviews.to_csv(ruta_reviews_csv, index=False, sep=';', encoding='utf-8')

df_users_reviews = pd.read_csv('users_reviews.csv',sep=";")

Removendo Duplicados

df_users_reviews = df_users_reviews.drop_duplicates()

guardando el archivo como parquet

df_users_reviews.to_parquet(ruta_parquet_reviews)

**tratamiento y limpieza de datos**

tratamiento y limpieza archivo steam_games

#lectura del archivo
df_games = pd.read_parquet(ruta_parquet_games)

df_games.head()

Primera línea no es necesária para el dataframe

df_games.drop(0, inplace=True)

df_games.head()

Normalizando la columna "price"

# Unifico todos los valores distintos hacia un unico valor 'Free To Play'
# Usa la función .replace() para realizar la transformación
df_games['price'] = df_games['price'].replace('Free to Play', 'Free To Play')
df_games['price'] = df_games['price'].replace('Free Demo', 'Free To Play')
df_games['price'] = df_games['price'].replace('Play for Free!', 'Free To Play')
df_games['price'] = df_games['price'].replace('Install Now', 'Free To Play')
df_games['price'] = df_games['price'].replace('Play WARMACHINE: Tactics Demo', 'Free To Play')
df_games['price'] = df_games['price'].replace('Free Mod', 'Free To Play')
df_games['price'] = df_games['price'].replace('Install Theme', 'Free To Play')
df_games['price'] = df_games['price'].replace('Third-party', 'Free To Play')
df_games['price'] = df_games['price'].replace('Play Now', 'Free To Play')
df_games['price'] = df_games['price'].replace('Free HITMAN™ Holiday Pack', 'Free To Play')
df_games['price'] = df_games['price'].replace('Play the Demo', 'Free To Play')
df_games['price'] = df_games['price'].replace('Free to Try', 'Free To Play')
df_games['price'] = df_games['price'].replace('Free to Use','Free To Play')
df_games['price'] = df_games['price'].replace('Free','Free To Play')

def replace_free_to_play(df):
    df["price"] = df["price"].replace("Free To Play", 0.0)
    return df


# Aplica la función para reemplazar 'Free To Play' por 0.0
df_games = replace_free_to_play(df_games)

# Limpia la columna "price" para quedarse solo con valores numéricos y decimales
df_games['price'] = df_games['price'].apply(lambda x: re.sub(r'[^\d.]', '', str(x)))
# Reemplaza cadenas vacías con NaN
df_games['price'].replace('', np.nan, inplace=True)
# Reemplazar valores nulos por 0.0 en la columna "price"
df_games["price"].fillna(0.0, inplace=True)
# Convierte la columna en tipo float
df_games['price'] = df_games['price'].astype(float)

df_games.info()

renombrando columna id para item_id

df_games = df_games.rename(columns = {"id":"item_id"})

df_games.dtypes

Conversión de la columna "item_id" de float a int

df_games['item_id'] = df_games['item_id'].astype(int)

mezclando columnas repetidas

"app_name" y "title"

df_games['title'].fillna(df_games['app_name'], inplace=True)

# Excluir a coluna "app_name"
df_games.drop('app_name', axis=1, inplace=True)

df_games.info()

"publisher" y "developer"

df_games['developer'].fillna(df_games['publisher'], inplace=True)

# Excluir a coluna "app_name"
df_games.drop('publisher', axis=1, inplace=True)

df_games.info()

Creación de la columna "year"

# Definir una función personalizada para extraer el año de la fecha
def extract_year(date_str):
    try:
        return int(date_str.split("-")[0])
    except (ValueError, AttributeError):
        return 2016  # Cambia 2016 al valor predeterminado que desees

# Aplicar la función personalizada a la columna "release_date"
df_games["year"] = df_games["release_date"].apply(extract_year)

# Convertir la columna "year" a tipo int
df_games["year"] = df_games["year"].astype(int)


df_games.head()

limpiando las columnas no necesárias restantes

df_games = df_games.drop(columns=["url","reviews_url","early_access","specs","tags","release_date","early_access"], axis=1)

df_games

df_games.info()

Eliminando valores nulos del dataframe

df_games = df_games.dropna(axis=0)

df_games.info()

df_games.dtypes

Conversión del dataframe limpio en parquet

df_games.to_parquet(ruta_parquet_games, index=False)

tratamiento y limpieza archivo data_reviews

df_reviews = pd.read_parquet(ruta_parquet_reviews)

df_reviews.head()

Normalización columna "posted"

def extraer_año(fecha):
    # Las fechas vienen en un formato del estilo 'Posted: July 15, 2013.'. Las fechas del año del dataset (2016) no
    # incluyen el año (por ejemplo, 'Posted: March 23.'). Usamos una expresión regular para extraer el año.
    result = re.search("\w* \d*, (\d*)", fecha)
    if result is not None:
        return int(result.group(1))
    else:
        return 2016

df_reviews["posted"] = df_reviews["posted"].apply(extraer_año)

df_reviews.head()

df_reviews.info()

df_reviews.isna().sum()

Conversión del dataframe limpio en parquet

df_reviews.to_parquet(ruta_parquet_reviews, index=False)

tratamiento y limpieza archivo data_items

df_items = pd.read_parquet(ruta_parquet_items)

df_items.head()

df_items.info()

df_items.dtypes

df_items.isna().sum()

Optimización del dataframe

df_items = df_items[:2000000]

df_items.info()

Conversión del dataframe limpio en parquet

df_items.to_parquet(ruta_parquet_items, index=False)

**Análisis de Sentimiento**

import json
import pandas as pd
import ast
import re
from textblob import TextBlob

df_reviews = pd.read_parquet(ruta_parquet_reviews)

df_reviews.head()

# Transformando tipo de dato float a str para la columna 'review' debido a que el análisis de sentimiento arrojaba un error por el tipo de dato
df_reviews['review'] = df_reviews['review'].astype(str)

# Realizamos el análisis de sentimiento para la columna 'review' de df_reviews

from textblob import TextBlob

# Función para analizar el sentimiento y asignar valores según la escala
def analizar_sentimiento(texto):
    analysis = TextBlob(texto)
    if analysis.sentiment.polarity < 0:
        return 0  # Sentimiento negativo
    elif analysis.sentiment.polarity == 0:
        return 1  # Sentimiento neutral
    else:
        return 2  # Sentimiento positivo

# Aplicar la función de análisis de sentimiento a la columna 'texto' y crear 'sentiment_analysis'
df_reviews['sentiment_analysis'] = df_reviews['review'].apply(analizar_sentimiento)

df_reviews.head()

df_reviews = df_reviews.drop("review", axis=1)

df_reviews.head()

Guardando cambios del DataFrame como parquet

df_reviews.to_parquet(ruta_parquet_reviews, index=False)
